requirements.txt
email
smtplib
pandas 
concurrent.futures
time
dotenv
os
google.generativeai